move objects over to helperfunctions.ts
