+++ src/components/TableFilters/TableFilters.tsx
@@ -7,7 +7,7 @@ import CloseIcon from '@mui/icons-material/Close';
 import RefreshIcon from '@mui/icons-material/Refresh';
 import Draggable from 'react-draggable';
 import ApprovalIcon from '@mui/icons-material/Approval';
-import { DEFAULT_FILTER_OPTIONS } from '../../assets/filterOptions';
+import { DEFAULT_FILTER_OPTIONS } from '../../assets/helperFunctions';
 import { FilterOption, TableFiltersProps } from '../../types';
 
 const TableFilters = ({ applyFilters, appliedFilters, filter_options }: TableFiltersProps) => {
